pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "This application is a page analyzer. Made as part of the Python Back End Development course."
authors = ["nickolay_kuzmin <nickolay.kuzmin2001@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
flake8 = "^6.0.0"
gunicorn = "^20.1.0"
validators = "^0.20.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
